+++ lofty/src/ogg/opus/properties.rs
@@ -21,6 +21,7 @@ pub struct OpusProperties {
 	pub(crate) channel_mask: ChannelMask,
 	pub(crate) version: u8,
 	pub(crate) input_sample_rate: u32,
+	pub(crate) output_gain: i16,
 }
 
 impl From<OpusProperties> for FileProperties {
@@ -76,6 +77,11 @@ impl OpusProperties {
 	pub fn input_sample_rate(&self) -> u32 {
 		self.input_sample_rate
 	}
+
+	/// Output gain in dB
+	pub fn output_gain_db(&self) -> f32 {
+		f32::from(self.output_gain) / 256.0
+	}
 }
 
 pub(in crate::ogg) fn read_properties<R>(
@@ -101,8 +107,7 @@ where
 	let pre_skip = identification_packet_reader.read_u16::<LittleEndian>()?;
 
 	properties.input_sample_rate = identification_packet_reader.read_u32::<LittleEndian>()?;
-
-	let _output_gain = identification_packet_reader.read_u16::<LittleEndian>()?;
+	properties.output_gain = identification_packet_reader.read_i16::<LittleEndian>()?;
 
 	let channel_mapping_family = identification_packet_reader.read_u8()?;